# Increase scan performance by another ~30% by passing each scan item