directly to JS through an ArrayBuffer as soon as we have it.

replicache_client.js: 24484
replicache_client.js.br: 4601
replicache_client_bg.wasm: 516997
replicache_client_bg.wasm.br: 149987

Running benchmarks please wait...
populate 1024x1000 (clean) x 2.22 MB/s ±0.0% (0 runs sampled)
populate 1024x1000 (dirty) x 1.66 MB/s ±0.0% (0 runs sampled)
scan 1024x1000 x 25.70 MB/s ±0.0% (0 runs sampled)
scan 1024x5000 x 29.41 MB/s ±0.0% (0 runs sampled)